requirements.txt
openai
pandas
streamlit
numpy 
python-dotenv